requirements.txt
numpy
pandas
pillow
tqdm
torchio
opencv-python
tqdm